lambda/tools/image-analysis/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in --output-file requirements.txt
annotated-types==0.7.0
    # via pydantic
aws-lambda-powertools==3.5.0
    # via -r requirements.in
aws-xray-sdk==2.14.0
    # via -r requirements.in
botocore==1.36.17
    # via aws-xray-sdk
cachetools==5.5.1
    # via google-auth
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
google-auth==2.38.0
    # via google-genai
google-genai==0.6.0
    # via -r requirements.in
idna==3.10
    # via requests
jmespath==1.0.1
    # via
    #   aws-lambda-powertools
    #   botocore
pillow==10.2.0
    # via google-genai
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.10.6
    # via google-genai
pydantic-core==2.27.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via botocore
requests==2.32.3
    # via google-genai
rsa==4.9
    # via google-auth
six==1.17.0
    # via python-dateutil
typing-extensions==4.12.2
    # via
    #   aws-lambda-powertools
    #   pydantic
    #   pydantic-core
urllib3==2.3.0
    # via
    #   botocore
    #   requests
websockets==14.2
    # via google-genai
wrapt==1.17.2
    # via aws-xray-sdk
